create tower shoot

+++ Assets/Scripts/Towers/TowerConfig.cs
@@ -17,7 +17,7 @@ public class TowerLevel
     [SerializeField] private int _cost;
     [SerializeField] private int _dmg;
     [SerializeField] private float _range;
-    [SerializeField] private double _fireRate;
+    [SerializeField] private float _fireRate;
     [SerializeField] private int _levelIndex;
     [SerializeField] private Mesh _buildingMesh;
     [SerializeField] private Mesh _finalMesh;
@@ -25,7 +25,7 @@ public class TowerLevel
     public int Cost { get => _cost; set => _cost = value; }
     public int Dmg { get => _dmg; set => _dmg = value; }
     public float Range { get => _range; set => _range = value; }
-    public double FireRate { get => _fireRate; set => _fireRate = value; }
+    public float FireRate { get => _fireRate; set => _fireRate = value; }
     public int LevelIndex { get => _levelIndex; set => _levelIndex = value; }
     public Mesh BuildingMesh { get => _buildingMesh; set => _buildingMesh = value; }
     public Mesh FinalMesh { get => _finalMesh; set => _finalMesh = value; }
